Add support for moving cursor within a row
CSI 'G' command

Re-use the existing SetCursorCommand action, but allow both x and y to
be optional. In the none case we do not move the cursor on that axis.
This allows the G command to be implemented as it requests that the
cursor moves along X, but not Y