Ensure unknown CSI sequences are still fully parsed
In the existing code, when we see a csi sequence we do not recognize, we
bail early and stop parsing. This is incorrect, as there is a defined
format that all CSIs follow, and just because we don't understand it,
doesn't mean that the garbage from the sequence should be printed to the
screen

A better behavior is to log the failure, but do our best to continue on
without doing what was requested of us

Update CSI parsing to continue to parse, even if invalid. Generalize the
parsing to follow the EMCA-48 specification