Fix carriage return
After updating our terminfo, bash will now try to insert a carriage
return on long lines. Fix the behavior of \r to put the cursor at the
beginning of the line

This results in new problematic behavior though, when we see \r\n, we
should not insert a newline at the start of the line. Instead we should
interpret the \n as an instruction to move the cursor down, and inject a
newline at the _end_ of the line if necessary. This means we have to
manually handle newlines as well