Add scrollback support
Split existing TerminalBuffer into two. The visible portion indicates
what is currently on screen, and the viewport where cursor moves should
be relative to. The scrollback is everything else.

Data model is essentially the same, just when we convert cursor to
buffer position we only take the visible range into account

When retrieving data, we need to adjust formatting tags from absolute
positions to the split positions as well

Current UI displays scrollback and canvas as two separate labels, which
makes calculating the position of the cursor easier, this is a limited
solution (e.g. selecting text across the scrollback boundary)

Bonus fix: Fix TerminalBuffer not using the correct w/h