Add control key support
It looks like the linux line discipline is in charge of handling the
special keycodes. E.g. Ctrl-\ -> SIGQUIT, Ctrl-c -> SIGINT, Ctrl-z ->
SIGTSTP are all handled in n_tty.c

In order for us to get expected interactive behavior (being able to
stop applications, etc.), all we have to do is send the appropriate
control codes to the pty