Implement home and end keys
Pipe home and end keys from the GUI down to the terminal emulator. From
experimenting on other terminals, it seems that home and end follow the
same behavior as the arrow keys in decckm mode. I did not see any
documentation on this, however testing in konsole it seems to be
consistent

Needs Testing on Windows